backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
openai
python-multipart
websockets==12.0
cryptography
mysql-connector-python
google-generativeai
